source/strafer_vlm/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "strafer-vlm"
version = "0.1.0"
description = "Workstation-first VLM utilities for Strafer (Qwen2.5-VL grounding)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pillow",
]

[project.optional-dependencies]
qwen = [
    "accelerate",
    "datasets",
    "peft",
    "torch",
    "transformers",
    "trl",
]
live = [
    "opencv-python",
]
service = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
]
dev = [
    "pytest>=7.0",
    "httpx",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["strafer_vlm*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: marks tests that require a GPU and loaded model (deselect with '-m \"not gpu\"')",
]
